pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "stvc"
version = "0.1.0"
description = "Speech-to-Text for Vibe Coding — system-wide voice dictation for Windows"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "faster-whisper",
    "sounddevice",
    "pywin32",
    "pynput",
    "pystray",
    "Pillow",
]

[project.scripts]
stvc = "stvc.app:main"

[tool.setuptools.packages.find]
where = ["src"]
